feat: 이미 존재하는 시간에 대한 예약이 존재할 때 삭제 불가 예외 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,17 +2,22 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.dto.ReservationTimeResponseDto;
+import roomescape.exception.ReservationTimeException;
 
 @Service
 public class ReservationTimeService {
 
+    private final ReservationDao reservationDao;
     private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationDao reservationDao,
+        ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
         this.reservationTimeDao = reservationTimeDao;
     }
 
@@ -30,6 +35,9 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        if (reservationDao.findByTimeId(id) != 0) {
+            throw new ReservationTimeException("이미 예약된 예약 시간을 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteReservationTime(id);
     }
 }
